requirements.txt
gradio>=6.5.1
httpx>=0.28.1
python-dotenv>=1.2.1
mcp[cli]>=1.26.0
caldav>=1.3.6
vobject>=0.9.9
google-generativeai>=0.8.3
pydantic>=2.10.6
